Update DSpaceMetsAssembler to obtain the DspaceMetsPackageProviderFactory when the PackageStream is created, instead of when the DSpaceMetsAssembler is constructed.
Now that there are multiple instances of DspaceMetsPackageProviderFactory in the Spring ApplicationContext, the package provider factory cannot be injected when constructing the DspaceMetsAssembler.  Instead the assembler will ask the concrete subclass for the package provider factory.

Because the assembler implementation is now responsible for return the package provider factory, the DspaceMetsPackageProviderFactory no longer needs to be registered as a Spring @Component.